+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.8] - 2018-03-11
+
+### Fixed
+- Fix first block interpolation
+- Fix mark out algorithm
+- Fix mark history files
+
+### Changed
+- Markout files prefix to underscore
+- Change graph A to blue
+
 ## [0.2.7] - 2018-03-04
 
 ### Fixed